requirements.txt
# ARKit to COLMAP converter dependencies
numpy>=1.24.0
scipy>=1.10.0
pillow>=9.0.0
tqdm>=4.65.0

# For cloud upload (optional)
boto3>=1.28.0
google-cloud-storage>=2.10.0
